\name{SAIC-package}
\alias{SAIC-package}
\alias{SAIC}
\docType{package}
\title{
\packageTitle{SAIC}
}
\description{
\packageDescription{SAIC}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SAIC}
\packageIndices{SAIC}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{SAIC}

Maintainer: \packageMaintainer{SAIC}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
